feat(tools): rename utils.py to api_utils.py and update imports
- Rename utils.py to api_utils.py for better clarity of its purpose
- Update __init__.py to expose all utility functions from api_utils
- Adjust import statements in dependent scripts to use the new module structure
- Maintains all existing functionality while improving code organization
