+++ docs/guides/model_convert/convert_from_pytorch/tools/utils/__init__.py
@@ -1,7 +1,19 @@
 from .api_url_parser import get_parser
-from .utils import get_url
+from .api_utils import (
+    escape_underscores_in_api,
+    extract_no_need_convert_list,
+    get_base_dir,
+    get_url,
+    load_mapping_json,
+    parse_md_files,
+)
 
 __all__ = [
     "get_parser",
     "get_url",
+    "escape_underscores_in_api",
+    "load_mapping_json",
+    "parse_md_files",
+    "get_base_dir",
+    "extract_no_need_convert_list",
 ]
